2.project-root-ai-evaluation/backend/requirements.txt
# FastAPI 核心依赖
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis 和任务队列
redis==5.0.1
celery==5.3.6

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 图像处理
opencv-python==4.9.0.80
Pillow>=10.3.0
numpy==1.26.3

# OCR 集成
baidu-aip==4.16.13
tencentcloud-sdk-python==3.0.1074

# 对象存储
oss2==2.18.4

# HTTP 客户端
httpx==0.26.0
aiohttp==3.9.1

# PDF 生成
weasyprint==60.2
playwright==1.41.0

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
hypothesis==6.98.1

# 监控和日志
prometheus-client==0.19.0
python-json-logger==2.0.7

# 工具
python-dotenv==1.0.0
